// src/features/coordinateurs/models/coordinateur.ts
export type Coordinateur = {
  id: string;
  prenom: string;
  nom: string;
  commune: string;
  codePostal: string;
  adresse: string;
  courriel: string;
  telephone: string;
  perimetre: string;
  nombreDePersonnesCoordonnees: number;
  nombreDeStructuresAvecDesPersonnesCoordonnees: number;
  dispositif: string;
  latitude: number;
  longitude: number;
};